result merge fix
