+++ src/features/Results/CaseResult/CaseResultView/CaseResultView.styled.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import { spacings } from '../../../tokens/spacings';
-export const CaseResults = styled.div`
+import { spacings } from '../../../../tokens/spacings';
+export const CaseResultView = styled.div`
   display: flex;
   flex-direction: column;
